backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pyjwt==2.12.1
python-dotenv==1.2.2
razorpay==2.0.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
